+++ var/starter_kits/italiagov/src/js/custom/custom.js
@@ -8,3 +8,25 @@
 //import './example-alert-event'
 //import './example-bootstrap-italia-modules'
 //import './icons'
+
+/**
+ * Customize your bootstrap
+ */
+
+// Example: add your custom object
+/*
+ bootstrap.CustomObject = CustomObject;
+*/
+
+// Example: remove unused object
+/*
+ if ('Object' in bootstrap) {
+   bootstrap.Object = undefined;
+ }
+ */
+
+// Alternatively, instead of adding or removing objects,
+// you can rebuild "bootstrap" with only the objects you want to use.
+/*
+ const { A, B, C } = bootstrap;
+ */